Remove "maxConnections" config
Turns out, that's not what ServerSocket's backlog property does. It's
the kernel's backlog of connections waiting to be `accept`ed by the
server. Instead, just setting using the default `backlog`.

However, this means we're also not tieing the # of threads in the pool
to the max connections. Instead, we're using the number of
"availableProcessors" according to the runtime.