Refactor line parsing into separate class, prevent stream consumption
Turns out request line reading was also consuming the stream :-/. Now we
pass the body as a lower-level input stream, which allows reading plain
bytes.

Getting the property-based test to pass was tough, but the key finding
was to serialize the body string's *code points* and to be sure not to
accidentally truncate them when serializing the entire request.