Refactor TcpServer to decouple from Socket, test err handling
Introduces the SocketController (and related) interfaces to act as a
"seam" between the TcpServer and the underlying client & server sockets.
This allows the TcpServer class to focus on the high level
responsibilities it owns:

- Binding to a port
- Passing new connections to its handler
- Handling errors gracefully at various points

The "concrete" AsyncServerSocketController then has the following
responsibilities passed on to it:

- Offloading connection acceptance & handling to other threads
- Configuring client sockets (read timeout)

All of the above are tested independently, and then tested together in
the pre-existing integration & acceptance tests.