Refactor request & response builder/serialization. Fix response bug.
They're identical with the exception of the request/status line, makes
sense for them to share code. As they're both HTTP messages, they both
depend on the classes in the `message` package.

For now, the refactor preserves API compatibility with RequestReader and
so on. A future refactor might change this.

This refactor also surfaced a bug which caused response bodies to not be
sent properly when there weren't any headers. This shouldn't be a
problem in practice (since responses with a body require a
Content-Length and other headers), but good to catch for spec
conformance.

Also, in debugging this I happened upon a case where the response
inequality property-based test failed due to hash inequality not
matching object inequality 1:1. I removed the assertion since it's
possible for unequal objects to have the same hash (but not vice versa).